pyproject.toml
[project]
name = "mcp-ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.2.1",
    "langchain-tavily>=0.2.18",
    "mcp[cli]>=1.27.0",
    "python-dotenv>=1.2.2",
]
